+++ src/test/java/it/univaq/cdvd/test_logica_business/RegistrazioneControllerTest.java
@@ -4,6 +4,8 @@ import it.univaq.cdvd.controller.RegistrazioneController;
 import it.univaq.cdvd.dao.UtenteDAO;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 public class RegistrazioneControllerTest {
@@ -12,14 +14,14 @@ public class RegistrazioneControllerTest {
     @Test
     public void testUnvalidRegistration() {
         UtenteDAO utenteDAO = new UtenteDAO();
-        assertFalse( utenteDAO.saveUser("", "", "",0.00));
+        assertFalse( utenteDAO.saveUser("", "", "",BigDecimal.valueOf(0)));
     }
 
     // test che verifica che non si possano inserire utenti già esistenti
     @Test
     public void testExistingUser() {
         UtenteDAO utenteDAO = new UtenteDAO();
-        assertFalse(utenteDAO.saveUser("prova", "x", "x",0));
+        assertFalse(utenteDAO.saveUser("prova", "x", "x", BigDecimal.valueOf(0)));
 
     }
 }
